chore: renamed files in chat to page
